重構寫法
