+++ src/components/CCC/PageHeader.vue
@@ -1,0 +1,61 @@
+<template>
+  <div class="page-header">
+    <h1>AI 智慧推薦</h1>
+  </div>
+</template>
+
+<style scoped>
+/* 頁面標題和搜尋區 */
+.page-header {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin-bottom: 20px;
+  flex-wrap: wrap;
+  gap: 15px;
+}
+
+h1 {
+  font-size: 24px;
+  margin: 0;
+  color: #2e7d32;
+}
+
+.search-bar {
+  display: flex;
+  align-items: center;
+  min-width: 300px;
+}
+
+.search-bar input {
+  height: 36px;
+  padding: 0 10px;
+  border: 1px solid #ddd;
+  border-radius: 4px 0 0 4px;
+  flex-grow: 1;
+  outline: none;
+}
+
+.search-btn {
+  background-color: #2e7d32;
+  color: white;
+  border: none;
+  border-radius: 0 4px 4px 0;
+  padding: 8px 16px;
+  cursor: pointer;
+  height: 36px;
+}
+
+/* 手機版適配 */
+@media (max-width: 768px) {
+  .page-header {
+    flex-direction: column;
+    align-items: flex-start;
+  }
+
+  .search-bar {
+    width: 100%;
+    min-width: unset;
+  }
+}
+</style>
